Convert Game/src/item/KinkyDungeonItem
I'm pretty sure `KinkyDungeonItemEvent()` takes an optional boolean as
its second argument.  However, there's one call site that was passing in
either a KDRestraintVariant or null if the lookup failed.  I suspect it
was type-punning it to a boolean, so I made it explicit at the call
site.  I am not completely sure this was correct.