Convert Game/src/enemy/KDAIList
Changed type signatures for functions in `AIType` -- the `aidata`
parameter is now of type `KDAIData`.

Changed the name of parameters from `AIData` to `aiData`.  This
parameter was rarely used, and would normally get prefixed with an
I-meant-to-not-use-that underscore.  But there's also a global variable
named `AIData`.  If you added any new code attempting to reference the
parameter `AIData`, you might instead reference the global `AIData`,
which would be a mess and *very* difficult to track down.  So the
parameters got renamed to `aiData` to avoid that.