Convert Game/src/enemy/KinkyDungeonEnemyEventList
Once again, we had function parameters named `AIData`; they've all been
renamed `aiData`.

***However!***  The `maintain()` method type signature specifies
`aiData` as an optional third parameter.  Some instances of `maintain()`
didn't include the `aiData` parameter, but *did* access a variable in
the function body named `AIData`.  Question:  Did this code intend to
access the missing parameter, or did it intend to access the global?

I have assumed they all intended to access the unspecified parameter,
and adjusted all `maintain()` instances to reflect this.  But this
almost certainly will result in a behavioral change.  (And none of the
instances, new or pre-existing, check if `aiData` is null/undefined.)