Convert Game/src/enemy/KinkyDungeonEnemyTeaseAttacks
Again, issues with parameters named `AIData`; all renamed to `aiData` to
avoid conflict with the global `AIData`.

Also: In the `apply()` instances, the field `dmg.string` was accessed
from the `dmg` variable, but `dmg` would be created either as an object
literal with the properties `{ dmg; happened; }` or returned from
`KinkyDungeonDealDamage()` as `{ string; happened; }`.  All the literals
have been fixed to `{ string; happened; }`.