Fix hotbar hotkey "hang".
The hotbar is loaded from localStorage from the storage keys
`KinkyDungeonSpellsChoice[123]`.  However, if these localStorage keys
don't exist, then the hotkeys to switch directly to them stop working.
This was because the test for storage key existence was made for the
previous index value, not the new candidate.  The invalid new value
would be stored in `KinkyDungeonSpellsConfig`, and the localStorage test
would fail forevermore after that.

Test for existence of the new storage key value.  If present, assign
KinkyDungeonSpellsConfig to the new index and update the hotbar.  If the
storage key does not exist, `KinkyDungeonSpellsConfig` and the live
hotbar will remain unchanged.

Also: Fix a wrong index when configuring the hotkey buttons.

Also also: Apply the correct type to `list` when loading the spells
config (bearing in mind `JSON.parse()` does no schema validation).