Reworked Mod Loading (again)
Using only promises may have caused a bit of a headache to understand it and also did not deal with the eval problem. Rework now awaits the result of eval and promisifies it. This appears to work in my testing as well as being significantly easier to follow and read.