Make Gold Duration Variable
This change allows overwriting the two floor timer of the gold lock with a variable timer as defined in the restraint data. 'lockDuration' will be called to determine how many floors must be passed. It will appear in the restraint data as: 'lockDuration: x' where x is the number defining how many floors must be cleared. If there is no lock duration entry within the restraint data, the duration will default to 2, thereby not breaking the behavior of existing gold locks.