Titles and Achievements
Implements a new tab under the Journal section which depicts milestones the player can undertake, as well as equipping one of them as a title to display under their name.

These milestones range from simple (play as a class) to longer term (reach NG+) and should persist in LocalStorage across sessions. Code is also present to implement effects when the title is equipped or unequipped, however this needs iteration.

Also implements an event that is called when bondage dialogues end to set the sub or dom flag.

This does have a few unfortunate ts-ignores.