Fix random ranges.
Entries in `KinkyDungeonLootTable[]` contain the fields:

	minEnchants
	maxEnchants
	minHex
	maxHex

At a glance, the ranges appear inclusive.  However, the calculation in
`KinkyDungeonLootEvent()` fails to take into account that `KDRandom()`
returns a value in the range [ 0.0, 1.0 ), which means that, after
`Math.floor()`, the upper value of the range would never be selected.

Add a fencepost to the calculation of quantities to ensure the full
intended range can be selected.

(Note: `minHex` and `maxHex` appear in this change, but I can find no
items that use those fields.)